Assignment 7
Assignment 7

Create a simple webpage that allows users to add tasks to a to-do list. Each task should have
a title and description. Users should be able to mark tasks as completed, edit task details,
and remove tasks from the list.
Webpage that includes:
A heading (&lt;h1&gt;) with the title &quot;To-Do List&quot;.
An input form that allows users to enter a task title and description.
A button labeled &quot;Add Task&quot; that will add the task to the list when clicked.
When the &quot;Add Task&quot; button is clicked:
Create a new div for each task.
Inside this div, display the task title and description.
Add a &quot;Mark as Completed&quot; button, &quot;Edit&quot; button, and &quot;Delete&quot; button for each task.
When the &quot;Mark as Completed&quot; button is clicked:
Add a class called &quot;completed&quot; to the task&#39;s title and description to apply a line-through
effect.
Update the button text to &quot;Mark as Incomplete&quot; and toggle the task&#39;s completion status.
When the &quot;Delete&quot; button is clicked:
Remove the task div from the taskList.
When the &quot;Edit&quot; button is clicked:
Allow the user to edit the task title and description. Replace the static text with input fields
pre-filled with the current task data.
After editing, the button text should change to &quot;Save&quot;, and clicking &quot;Save&quot; should update
the task with the new details.